// src/parser/types.ts
export interface Attributes {
  [key: string]: Primitive | { [key: string]: Primitive }
}

export interface Attributed {
  attributes: Attributes
}
// ...
type TableContent = TableRow | TableRule

export interface Table extends Parent, Attributed {
  type: 'table'
  children: TableContent[]
}

export interface TableRow extends Parent {
  type: 'table.row'
  children: TableCell[]
}
// ...
export interface TableRule extends Node {
  type: 'table.hr'
}